+++ src/interface/pwm.h
@@ -37,7 +37,7 @@ struct Pwm
     };
 
     virtual bool open() = 0;
-    virtual bool set(int val, const string& sourceName) = 0;
+    virtual bool set(double val, const string& sourceName) = 0;
     virtual void setMin(int val) = 0;
     virtual void setMax(int val) = 0;
     virtual void setMode(Mode mode) = 0;
